+++ examples/getting-started/01-prompting.js
@@ -1,2 +1,2 @@
-var cleave = require('cleave');
-cleave.prompt('How are you?');
+var climate = require('climate');
+climate.prompt('How are you?');