+++ examples/getting-started/02-simple-receive.js
@@ -1,6 +1,6 @@
-var cleave = require('cleave');
-cleave.prompt('How are you?');
+var climate = require('climate');
+climate.prompt('How are you?');
 
-cleave.receive('well', function() {
-    cleave.out('Great to see you are well\n');
+climate.receive('well', function() {
+    climate.out('Great to see you are well\n');
 });