+++ examples/getting-started/03-wildcard-receive.js
@@ -1,6 +1,6 @@
-var cleave = require('cleave');
-cleave.prompt('How are you?');
+var climate = require('climate');
+climate.prompt('How are you?');
 
-cleave.receive('*', function(input) {
-    cleave.out('I see you are ' + input + '\n');
+climate.receive('*', function(input) {
+    climate.out('I see you are ' + input + '\n');
 });